+++ Sources/CoreMindTests/DeviceInfoTests.swift
@@ -15,7 +15,11 @@ final class DeviceInfoTests: XCTestCase {
 	}
 
 	func testModel() {
-		XCTAssertEqual(sut.model, "x86_64")
+        #if arch(arm64)
+		XCTAssertEqual(sut.model, "arm64")
+        #else
+        XCTAssertEqual(sut.model, "x86_64")
+        #endif
 	}
 
 	func testOs() {