+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
-## [Unreleased]
+## [Unreleased] 
+- Remove Result<T> type alias to Swift.Result<T, Error> (#12)
 
 ## 1.2.1 - 2020-11-07